+++ src/index.ts
@@ -2,4 +2,5 @@ export { Client, ClientError } from "./client";
 export type { ClientOptions } from "./client";
 export { Server } from "./server";
 export type { ServerHandler, ServerOptions } from "./server";
+export { sseEventsToMessages, resolvePendingToolUse } from "./utils";
 export type * from "./types";